+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
-## Unreleased
+## [0.9.2] - 2024-03-17
 
 ### Fixed
 - fixed automatic hiding of stone strata variants (bug introduced in 0.9.0)
@@ -128,6 +128,7 @@ Initial 1.20.1 release!
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
+[0.9.2]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.20.1-0.9.2
 [0.9.1]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.20.1-0.9.1
 [0.9.0]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.20.1-0.9.0
 [0.8.1]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.20.1-0.8.1